Added optimization to avoid calling the API multiple times uneccisarily

+++ pages/async-atom-effects/atoms-class.ts
@@ -54,10 +54,13 @@ export const itemState = atomFamily<ItemType, number>({
   key: 'item',
   default: { label: '', checked: false },
   effects_UNSTABLE: (id) => [
-    ({ onSet, setSelf }) => {
+    ({ onSet, setSelf, trigger }) => {
       setSelf(cachedAPI.getItem(id));
 
-      onSet((item) => {
+      onSet((item, oldItem) => {
+        // optimization to avoid calling the API multiple times uneccisarily
+        if (oldItem instanceof DefaultValue && trigger === 'get') return;
+
         if (item instanceof DefaultValue) {
           shoppingListAPI.deleteItem(id);
         } else {
